services/recipe-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
PyJWT==2.8.0
openai==1.12.0
python-dotenv==1.0.0
httpx==0.27.2